backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.10
aiofiles==24.1.0
google-generativeai==0.8.3
requests==2.32.5
psutil==5.9.8
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
